+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 2.2.7.0 (July 07, 2016)
 
 * Upgrade to Apache Cassandra 2.2.7
+* Fix mapper referenced by alias in sortFields
 
 ## 2.2.6.2 (July 05, 2016)